[compat, modify] Use size_t (unsigned type) insted of int.

+++ src/lib/coil/common/coil/Properties.cpp
@@ -432,9 +432,9 @@ namespace coil
    * @brief Get the number of properties
    * @endif
    */
-  int Properties::size() const
+  size_t Properties::size() const
   {
-    return static_cast<int>(propertyNames().size());
+    return propertyNames().size();
   }
 
   /*!
@@ -753,7 +753,7 @@ namespace coil
    * @endif
    */
   std::ostream&
-  Properties::_dump(std::ostream& out, const Properties& curr, int index)
+  Properties::_dump(std::ostream& out, const Properties& curr, size_t index)
   {
     if (index != 0) out << indent(index) << "- " << curr.name;
     if (curr.leaf.empty())
@@ -783,10 +783,10 @@ namespace coil
    * @brief Create indents
    * @endif
    */
-  std::string Properties::indent(int index)
+  std::string Properties::indent(size_t index)
   {
     std::string space = std::string();
-    for (int i(0); i < index - 1; ++i)
+    for (size_t i(0); i < index - 1; ++i)
       {
         space += "  ";
       }
@@ -800,10 +800,10 @@ namespace coil
    * @brief Create indents
    * @endif
    */
-  std::string indent(int index)
+  std::string indent(size_t index)
   {
     std::string space;
-    for (int i(0); i < index - 1; ++i)
+    for (size_t i(0); i < index - 1; ++i)
       {
         space += "  ";
       }
@@ -832,7 +832,7 @@ namespace coil
       return out;
   }
 
-  void Properties::_dump(std::string& out, const Properties& curr, int index) const
+  void Properties::_dump(std::string& out, const Properties& curr, size_t index) const
   {
       if (index != 0) out += indent(index) + "- " + curr.name;
       if (curr.leaf.empty())
@@ -854,7 +854,7 @@ namespace coil
       }
   }
 
-  void Properties::_dump(std::vector<std::string>& out, const Properties& curr, int index) const
+  void Properties::_dump(std::vector<std::string>& out, const Properties& curr, size_t index) const
   {
       if (index != 0) out.push_back(indent(index) + "- " + curr.name);
       if (curr.leaf.empty())
